Nuklear UI Themes
Added NuklearTheme abstract class for defining UI theme styles.
Refactored existing UI classes to reduce duplicate style code and use
proper stack changes for temporary style swaps.

+++ base/src/main/java/com/avogine/game/ui/nuklear/audio/AudioConfigUI.java
@@ -95,7 +95,7 @@ public class AudioConfigUI {
 	}
 	
 	private float getALVolume() {
-		return ((float) volume[0]) / 100.0f;
+		return volume[0] / 100.0f;
 	}
 	
 	private void setUIVolume(float alVolume) {
